Resized and retraced.
The board was smaller than it had to be; why/how I did that, I do not know.

Now, the board is close to the max dimension for low-cost fab. Wiring is
neater, because I knew what I was doing. So, power/ground route around
the edges (largely), and then I routed data.

It is still the case that all vias are drilled *under* the LEDs, which
maintains the appearance of the board.

I probably should move the header up (to create more space for the USB
cable), but... that would be a lot of re-routing at this point. And,
it might not be an issue.

SO. I will give it one more look, but will probably send this version off
for fab as "v2".